Add Child Controller. Add Child views. Fix Camp entity, CampController and Camps views.

+++ src/main/java/org/niebieskidom/bluecamps/controller/CampController.java
@@ -2,15 +2,15 @@ package org.niebieskidom.bluecamps.controller;
 
 import org.niebieskidom.bluecamps.entity.Camp;
 import org.niebieskidom.bluecamps.entity.Child;
+import org.niebieskidom.bluecamps.repositories.CampRepository;
+import org.niebieskidom.bluecamps.repositories.ChildRepository;
 import org.niebieskidom.bluecamps.services.CampService;
 import org.niebieskidom.bluecamps.services.ChildService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
@@ -18,18 +18,17 @@ import java.util.Optional;
 @Controller
 @RequestMapping("/camp")
 public class CampController {
-    private CampService campService;
-    private ChildService childService;
+    private final CampService campService;
+    private final ChildService childService;
+    private final CampRepository campRepository;
+    private final ChildRepository childRepository;
 
-    @Autowired
-    public CampController(CampService campService, ChildService childService) {
+
+    public CampController(CampService campService, ChildService childService, CampRepository campRepository, ChildRepository childRepository) {
         this.campService = campService;
         this.childService = childService;
-    }
-
-    @ModelAttribute("camps")
-    public List<Camp> camps() {
-        return campService.showCamps();
+        this.campRepository = campRepository;
+        this.childRepository = childRepository;
     }
 
 
@@ -37,7 +36,6 @@ public class CampController {
     public List<Child> children() {
         return childService.getChildren();
     }
-
 
     @GetMapping("/all")
     public String showAllCamps(Model model) {
@@ -52,52 +50,40 @@ public class CampController {
         return "camps/campForm";
     }
 
+
     @PostMapping("/add")
-    public String addCamp(@Valid Camp camp, BindingResult bindingResult, Model model) {
+    public String addCamp(@Valid Camp camp, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return "camps/campForm";
         } else {
             campService.addCamp(camp);
-            return "/camps/campList";
+            return "redirect:/camp/all";
         }
     }
 
-    @GetMapping("/edit/{id}")
-    public String editCamp(@PathVariable long id, Model model) {
-        Optional<Camp> optionalCamp = campService.getCamp(id);
-        Camp camp = optionalCamp.orElseThrow(() -> new EntityNotFoundException("Camp not found"));
+    @PostMapping("/edit")
+    public String editCamp(Model model, @RequestParam long id) {
+        Optional<Camp> camp = campService.getCamp(id);
         model.addAttribute("camp", camp);
         return "camps/campForm";
     }
-//    @GetMapping("/author/modify")
-//    public String updateAuthor(Author author) {
-//        authorDao.modifyAuthor(author);
-//        return "redirect:/author/list";
-//    }
-//
-//    @PostMapping ("/author/modify")
-//    public String updateAuthor(Model model, @RequestParam int id) {
-//        Author author = authorDao.getAuthorById(id);
-//        model.addAttribute("author", author);
-//        return "/author/form";
-//    }
 
 
-
-
-    @GetMapping("/delete/{id}")
-    public String deleteCamp(@PathVariable long id, Model model) {
-        model.addAttribute("camp", campService.getCamp(id));
-        return "/camps/deleteConfirmation";
+    @PostMapping("delete")
+    public String deleteCamp(Model model, @RequestParam long id) {
+        model.addAttribute("id", id);
+        return "camps/deleteConfirmation";
     }
 
-    @PostMapping("/delete/{id}")
-    public String confirmDeleteCamp(Model model, @PathVariable long id) {
-        //usunąć najpierw listę dzieci z tego obozu
-        model.addAttribute("camp", campService.getCamp(id));
+    @PostMapping("delete/confirmed")
+    public String executeDeleteCamp(@RequestParam Long id) {
+        Optional<Camp> optionalCamp = campService.getCamp(id);
+        Camp camp = optionalCamp.orElse(null);
+        if (camp ==null){                   //dodaj obsługę błędu
+            return "have problems";
+        }
         campService.deleteCamp(id);
-        return "/camps/campList";
+        return "redirect:/camp/all";
     }
-
 
 }
